Move hooks into library
