+++ packages/example/src/RecorderExample.js
@@ -1,10 +1,9 @@
-import { Recorder } from '@nabidreams/react-native-audio';
+import { Recorder, useRecorder } from '@nabidreams/react-native-audio';
 import React from 'react';
 import { Button, StyleSheet, Text, View } from 'react-native';
 
 import config from './config';
 import LevelBar from './LevelBar';
-import useRecorder from './useRecorder';
 
 const styles = StyleSheet.create({
   root: {
