+++ packages/react-native-audio/index.js
@@ -1,2 +1,4 @@
 export { default as Player } from './src/Player';
 export { default as Recorder } from './src/Recorder';
+export { default as usePlayer } from './src/usePlayer';
+export { default as useRecorder } from './src/useRecorder';
