+++ packages/react-native-audio/package.json
@@ -23,7 +23,7 @@
     "player"
   ],
   "peerDependencies": {
-    "react": "^16.8.1",
-    "react-native": ">=0.59.0-rc.0 <1.0.x"
+    "react": ">=16.8",
+    "react-native": ">=0.59"
   }
 }
