+++ packages/react-native-audio/src/usePlayer.js
@@ -1,10 +1,11 @@
-import { Recorder } from '@nabidreams/react-native-audio';
 import React from 'react';
 import { InteractionManager } from 'react-native';
 
+import Player from './Player';
+
 export default () => {
   const [state, setState] = React.useState();
-  const isStarted = state === Recorder.State.STARTED;
+  const isStarted = state == Player.State.STARTED;
 
   const start = React.useCallback(
     async (filePath) => {
@@ -12,22 +13,22 @@ export default () => {
         await stop();
       }
 
-      await Recorder.start(filePath);
+      await Player.start(filePath);
     },
     [isStarted],
   );
 
   const stop = React.useCallback(async () => {
-    await Recorder.stop();
+    await Player.stop();
   }, []);
 
   React.useEffect(function listenStateChange() {
     (async () => {
-      setState(await Recorder.getState());
+      setState(await Player.getState());
     })();
 
-    const subscription = Recorder.addListener(
-      Recorder.EventType.STATE_CHANGE,
+    const subscription = Player.addListener(
+      Player.EventType.STATE_CHANGE,
       ({ state }) => {
         setState(state);
       },
@@ -36,17 +37,17 @@ export default () => {
     return () => subscription.remove();
   }, []);
 
-  const [level, setLevel] = React.useState(Recorder.MIN_LEVEL);
+  const [level, setLevel] = React.useState(Player.MIN_LEVEL);
 
   React.useEffect(
     function handleLevelChange() {
       async function updateLevel() {
-        if ((await Recorder.getState()) !== Recorder.State.STARTED) {
-          setLevel(Recorder.MIN_LEVEL);
+        if ((await Player.getState()) !== Player.State.STARTED) {
+          setLevel(Player.MIN_LEVEL);
           return;
         }
 
-        setLevel(await Recorder.getLevel());
+        setLevel(await Player.getLevel());
 
         InteractionManager.runAfterInteractions({
           name: 'updateLevel',
@@ -64,7 +65,7 @@ export default () => {
 
   React.useEffect(() => {
     return () => {
-      Recorder.stop();
+      stop();
     };
   }, []);
