+++ packages/react-native-audio/src/useRecorder.js
@@ -1,10 +1,11 @@
-import { Player } from '@nabidreams/react-native-audio';
 import React from 'react';
 import { InteractionManager } from 'react-native';
 
+import Recorder from './Recorder';
+
 export default () => {
   const [state, setState] = React.useState();
-  const isStarted = state == Player.State.STARTED;
+  const isStarted = state === Recorder.State.STARTED;
 
   const start = React.useCallback(
     async (filePath) => {
@@ -12,22 +13,22 @@ export default () => {
         await stop();
       }
 
-      await Player.start(filePath);
+      await Recorder.start(filePath);
     },
     [isStarted],
   );
 
   const stop = React.useCallback(async () => {
-    await Player.stop();
+    await Recorder.stop();
   }, []);
 
   React.useEffect(function listenStateChange() {
     (async () => {
-      setState(await Player.getState());
+      setState(await Recorder.getState());
     })();
 
-    const subscription = Player.addListener(
-      Player.EventType.STATE_CHANGE,
+    const subscription = Recorder.addListener(
+      Recorder.EventType.STATE_CHANGE,
       ({ state }) => {
         setState(state);
       },
@@ -36,17 +37,17 @@ export default () => {
     return () => subscription.remove();
   }, []);
 
-  const [level, setLevel] = React.useState(Player.MIN_LEVEL);
+  const [level, setLevel] = React.useState(Recorder.MIN_LEVEL);
 
   React.useEffect(
     function handleLevelChange() {
       async function updateLevel() {
-        if ((await Player.getState()) !== Player.State.STARTED) {
-          setLevel(Player.MIN_LEVEL);
+        if ((await Recorder.getState()) !== Recorder.State.STARTED) {
+          setLevel(Recorder.MIN_LEVEL);
           return;
         }
 
-        setLevel(await Player.getLevel());
+        setLevel(await Recorder.getLevel());
 
         InteractionManager.runAfterInteractions({
           name: 'updateLevel',
@@ -64,7 +65,7 @@ export default () => {
 
   React.useEffect(() => {
     return () => {
-      stop();
+      Recorder.stop();
     };
   }, []);
